+++ NEWS.md
@@ -1,5 +1,6 @@
 # nexus 0.2.0.9000
 ## New classes and methods
+* Add `flatten()` to computes compositional mean of data subsets.
 * Add `show()` methods for `LogRatio-class` objects.
 
 # nexus 0.2.0